// src/api/endpoints.ts
export enum Endpoints {
  LOGIN = "/users/token/",
  LOGOUT = "/users/logout",
  PASSWORD_RECOVERY = "/users/reset_password/",
  FORGET_PASSOWRD = "/users/reset_password/set_password/",
  PASSWORD_TOKEN_VERIFY = "/users/reset_password/verify/",
  REFRESH_TOKEN = "/users/token/refresh/",
  GET_USER = "/users/me/",
  GET_USER_PERMISSIONS = "/users/permissions/",
  FETCH_USERS = "/users/",
  NOTARIES = "/notaries/",
  FETCH_OFFICES = "/admin/notaries/",
  FETCH_ROLES = "/admin/roles/",
  FETCH_PERMISSIONS = "/admin/roles/permissions/",
  CLIENTS = "/clients/",
  FETCH_ACTIONS = "/admin/actions/",
  APPLICATION = "/application/",
}